+++ Runtime/Parsers/ParserUtilities.cs
@@ -1,0 +1,8 @@
+﻿namespace InfiniteCanvas.InkIntegration.Parsers
+{
+	public static class ParserUtilities
+	{
+		public const string MULTI_VALUE_PARAMETER_DELIMITER = ";";
+		public const string PARAMETER_DELIMITER             = ">";
+	}
+}